# gh-93325: Micro-optimization of small integer lookup