--- a/Include/internal/pycore_long.h
+++ b/Include/internal/pycore_long.h
@@ -21,6 +21,7 @@ extern void _PyLong_FiniTypes(PyInterpreterState *interp);
 /* other API */
 
 #define _PyLong_SMALL_INTS _Py_SINGLETON(small_ints)
+#define _PyLong_SMALL_INTS_ZERO_OFFSET _Py_SINGLETON(small_ints_zero_offset)
 
 // _PyLong_GetZero() and _PyLong_GetOne() must always be available
 // _PyLong_FromUnsignedChar must always be available
@@ -31,16 +32,16 @@ extern void _PyLong_FiniTypes(PyInterpreterState *interp);
 // Return a borrowed reference to the zero singleton.
 // The function cannot return NULL.
 static inline PyObject* _PyLong_GetZero(void)
-{ return (PyObject *)&_PyLong_SMALL_INTS[_PY_NSMALLNEGINTS]; }
+{ return (PyObject *)&_PyLong_SMALL_INTS_ZERO_OFFSET[0]; }
 
 // Return a borrowed reference to the one singleton.
 // The function cannot return NULL.
 static inline PyObject* _PyLong_GetOne(void)
-{ return (PyObject *)&_PyLong_SMALL_INTS[_PY_NSMALLNEGINTS+1]; }
+{ return (PyObject *)&_PyLong_SMALL_INTS_ZERO_OFFSET[1]; }
 
 static inline PyObject* _PyLong_FromUnsignedChar(unsigned char i)
 {
-    return Py_NewRef((PyObject *)&_PyLong_SMALL_INTS[_PY_NSMALLNEGINTS+i]);
+    return Py_NewRef((PyObject *)&_PyLong_SMALL_INTS_ZERO_OFFSET[i]);
 }
 
 PyObject *_PyLong_Add(PyLongObject *left, PyLongObject *right);