--- a/Include/internal/pycore_runtime_init.h
+++ b/Include/internal/pycore_runtime_init.h
@@ -388,6 +388,7 @@ extern "C" {
             _PyLong_DIGIT_INIT(255), \
             _PyLong_DIGIT_INIT(256), \
         }, \
+        .small_ints_zero_offset = 0, \
         \
         .bytes_empty = _PyBytes_SIMPLE_INIT(0, 0), \
         .bytes_characters = { \