--- a/Misc/NEWS.d/next/Core and Builtins/2022-05-28-19-07-26.gh-issue-93325.6HUjfA.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-05-28-19-07-26.gh-issue-93325.6HUjfA.rst
@@ -1,0 +1,1 @@
+Micro optimization of small integer lookup.