The small integers are accessed through `_PyLong_SMALL_INTS`. This requires adding the constant `_PY_NSMALLNEGINTS` to the integer desired. This PR adds a new global variable `_PyLong_SMALL_INTS_ZERO_OFFSET` that points to the zero integer.

Advantages:

* Methods like `get_small_int`, `_PyLong_GetZero`, `_PyLong_GetOne` require an addition operation less.
* Since `_PyLong_GetZero` and `_PyLong_GetOne` are inline functions, less code is generated 

Benchmark:
```
Mean +- std dev: [main] 1.79 ms +- 0.10 ms -> [PR] 1.69 ms +- 0.18 ms: 1.06x faster
```

<details>
<summary>Test script</summary>

``` 
import pyperf
import time
runner = pyperf.Runner()

code=""" 
for ii in range(200):
    for jj in range(200):
        a = ii - jj
        b= ii+ jj
"""        

runner.timeit('small int test', stmt=code)
``` 

</details>


Notes:

* This PR adds a new singleton `small_ints_zero_offset`. This extra global can be avoided if we: first allocate `small_ints`, then add `_PY_NSMALLNEGINTS` to the pointer. For deallocating one first has to substract `_PY_NSMALLNEGINTS` again. This is a little more complex, but if preferred I will adapt the PR to this.
* I do not have sufficient knowledge of the Python initialization process to determine the best location to define and initialize the `small_ints_zero_offset` pointer. Suggestions for improvement are welcome.


<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
gh-NNNNN: Summary of the changes made
```

Where: gh-NNNNN refers to the GitHub issue number.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->

Fixes #93325.