@mdickinson 2022-05-29
Thanks for the PR.

How repeatable are those timings? A 6% speed up would be impressive, but it's well within the margin of error, and seems surprisingly large for what should be removal of a single assembly instruction (and possibly not even that - on Intel hardware, we could just be replacing an LEA instruction with an addition, keeping the total number of instructions the same).

As an experiment, I've been running your script in a loop, alternating between commit b8135e9f3c2854d4e4ead8df1342c20b5c2dabf7 on this branch and commit dec1e9346d82fa4a4761389c81d36ef9d01f332b on main. This is on macOS / Intel, with both branches configured and built with `./configure --enable-optimizations`. The numbers are quite variable, but so far I'm not seeing a convincing performance difference in either direction. Here's a sample:

```
main .....................
small int test: Mean +- std dev: 1.26 ms +- 0.03 ms

zero-offset .....................
small int test: Mean +- std dev: 1.25 ms +- 0.02 ms

main .....................
small int test: Mean +- std dev: 1.25 ms +- 0.03 ms

zero-offset .....................
small int test: Mean +- std dev: 1.26 ms +- 0.02 ms

main .....................
small int test: Mean +- std dev: 1.25 ms +- 0.02 ms

zero-offset .....................
small int test: Mean +- std dev: 1.25 ms +- 0.01 ms

main .....................
small int test: Mean +- std dev: 1.26 ms +- 0.03 ms

zero-offset .....................
small int test: Mean +- std dev: 1.26 ms +- 0.02 ms

main .....................
small int test: Mean +- std dev: 1.25 ms +- 0.02 ms

zero-offset .....................
small int test: Mean +- std dev: 1.27 ms +- 0.04 ms
```