@mdickinson 2022-05-29
Looking at the assembly on my machine, I do see a difference: here's what I typically see for the lookup on main (not surprisingly, `get_small_int` is inlined everywhere, so there are multiple versions of the code, but I've tried to pick typical versions):
```
	shlq	$5, %rax
	movq	__PyRuntime@GOTPCREL(%rip), %rcx
	leaq	(%rcx,%rax), %r14
	addq	$736, %r14                      ## imm = 0x2E0
```

And here's the corresponding code in this branch:
```
	movq	__PyRuntime@GOTPCREL(%rip), %rax
	movq	9120(%rax), %rax
	shlq	$5, %r15
	leaq	(%rax,%r15), %r14
```