@mdickinson 2022-05-29
Conclusion: I'm not yet seeing any evidence that this micro-optimization is actually an optimization. :-)

@markshannon Any thoughts on this PR?