@markshannon 2022-05-29
This is going to be slower.

It replaces an add with a load, which is almost always going to be slower, and never faster.