@sweeneyde 2022-05-29
In many situations, compilers may be able to constant-fold `small_ints[_PY_NSMALLNEGINTS + i]` as `(small_ints + _PY_NSMALLNEGINTS)[i]` anyway.