@eendebakpt 2022-05-29
@mdickinson @markshannon @sweeneyde Thanks for the responses. I did some more testing. The timings reported above are repeatable, but have been taken on a Windows system with the default build from the Visual Studio build files. I repeated the experiments, but now with binaries generated with the command `build.bat --pgo`. With this build configuration the running times are the same within random fluctuations. (not sure they are identical, but any improvement seems to small to further investigate)

I am still amazed that the compiler can handle this, but the numbers show no speedup so I will close this PR.